requirements.txt
timm>=0.4.12
torch>=1.10.1
torchvision>=0.11.2
pillow>=9.0.0
pytorch-lightning>=1.5.9
omegaconf>=2.1.1
hydra-core>=1.1.1
kornia>=0.6.2
tifffile>=2021.11.2
numpy>=1.21.2
segmentation-models-pytorch>=0.2.1
torchmetrics>=0.7.0
einops>=0.3.2
